refactor: isolate LocalDNS rendering
